Kits model pulled into Equipment / added to check program
