+++ server/wts/json/equipment/kits.json
@@ -5,7 +5,7 @@
     "code": "KA-1",
     "unitType": "Crisis",
     "cost": 2,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Other nations have implemented better preservation techniques, and have donated food to alleviate shortages",
     "effects": [{"type": "terror", "effect": -5 }],
     "prereq": [{ "type": "Technology", "code": "Agr2A"}],
@@ -17,7 +17,7 @@
     "code": "KA-2",
     "unitType": "Crisis",
     "cost": 3,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Better water supply techniques have been implemented, so water is more readily available than in the past.",
     "effects": [{"type": "terror", "effect": -5 }],
     "prereq": [{ "type": "Technology", "code": "Agr2B"}],
@@ -29,7 +29,7 @@
     "code": "KA-3",
     "unitType": "Clivilian",
     "cost": 2,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Farming techinques have been come easier to track and implement.",
     "effects": [{"type": "terror", "effect": -8 }],
     "prereq": [{ "type": "Technology", "code": "Agr3C"}],
@@ -41,7 +41,7 @@
     "code": "KA-4",
     "unitType": "Clivilian",
     "cost": 8,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Large food synthesizers to help feed this nation. Food insecurity has been greatly reduce as there are many of them around your nation.",
     "effects": [{"type": "terror", "effect": -10 }],
     "prereq": [{ "type": "Technology", "code": "Agr4C"}],
@@ -53,7 +53,7 @@
     "code": "KA-5",
     "unitType": "Crisis",
     "cost": 15,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Terraforming a distressed area in order to create a more beneficial environment.",
     "effects": [{"type": "terror", "effect": -15 }],
     "prereq": [{ "type": "Technology", "code": "Agr5A"}],
@@ -65,7 +65,7 @@
     "code": "KA-6",
     "unitType": "Clivilian",
     "cost": 25,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Individualized food synthesizers to help feed every citizen. Food insecurity is nearly a thing of the past in this nation.",
     "effects": [{"type": "terror", "effect": -15 }],
     "prereq": [{ "type": "Technology", "code": "Agr5B"}],
@@ -77,7 +77,7 @@
     "code": "KB-1",
     "unitType": "Crisis",
     "cost": 3,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Major breakthoughs in Anti-Radiation Treatments are now able to be deployed to Crisis.",
     "effects": [{"type": "terror", "effect": -5 }],
     "prereq": [{ "type": "Technology", "code": "Bio3A"}],
@@ -89,7 +89,7 @@
     "code": "KB-2",
     "unitType": "Base",
     "cost": 25,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Neuro-upgrades for your secret labs' scientists.",
     "effects": [{"sciRate": 400}],
     "prereq": [{ "type": "Technology", "code": "Bio4A"}],
@@ -101,7 +101,7 @@
     "code": "KB-3",
     "unitType": "Crisis",
     "cost": 25,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Robot nurses are sent into crisis. This allows for safer, easier, and quicker response, distributing medical care.",
     "effects": [{"type": "terror", "effect": -20 }],
     "prereq": [{ "type": "Technology", "code": "Bio5B"}],
@@ -113,7 +113,7 @@
     "code": "KB-4",
     "unitType": "Clivilian",
     "cost": 30,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Stable, always available Urgent Care centers have opened nation-wide, equipped with amazing robotic doctors with Baymax levels of diagnosing.",
     "effects": [{"type": "terror", "effect": -15 }],
     "prereq": [{ "type": "Technology", "code": "Bio5D"}],
@@ -125,7 +125,7 @@
     "code": "KI-1",
     "unitType": "Base",
     "cost": 25,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "AI computers increase computation speed.",
     "effects": [{"sciRate": 50}],
     "prereq": [{ "type": "Technology", "code": "Inf1D"}],
@@ -137,7 +137,7 @@
     "code": "KI-2",
     "unitType": "Clivilian",
     "cost": 10,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Materials deconstructed eats away at landfills.",
     "effects": [{"type": "terror", "effect": -5}, {"type": "$M", "effect": 1}],
     "prereq": [{ "type": "Technology", "code": "Inf2F"}],
@@ -149,7 +149,7 @@
     "code": "KI-3",
     "unitType": "Base",
     "cost": 25,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "AI computers increase computation speed.",
     "effects": [{"sciRate": 50}],
     "prereq": [{ "type": "Technology", "code": "Inf3D"}],
@@ -161,7 +161,7 @@
     "code": "KI-4",
     "unitType": "Crisis",
     "cost": 25,
-    "buildtime": 1,
+    "buildTime": 1,
     "desc": "Local weather patterns change in your favor. Unknown global effects.",
     "effects": [{"type": "terror", "effect": -40 }, {"type": "pr", "effect": -2 }],
     "prereq": [{ "type": "Technology", "code": "Inf4B"}],
@@ -173,7 +173,7 @@
     "code": "KI-5",
     "unitType": "Clivilian",
     "cost": 10,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Like the early Holodeck, these centers allow for life like interactions with digital enities. They also allow for those seperated by distance to have in-person interactions.",
     "effects": [{"type": "terror", "effect": -15}, {"type": "$M", "effect": -1}],
     "prereq": [{ "type": "Technology", "code": "Inf4D"}],
@@ -185,7 +185,7 @@
     "code": "KI-6",
     "unitType": "MegaProject",
     "cost": 30,
-    "buildtime": 3,
+    "buildTime": 3,
     "desc": "ARC-like cities, these structures hold entire large communities with every small ecological and physical footprint.",
     "effects": [{"type": "terror", "effect": -30}, {"type": "pr", "effect": 1}],
     "prereq": [{ "type": "Technology", "code": "Inf4F"}],
@@ -197,7 +197,7 @@
     "code": "KI-7",
     "unitType": "MegaProject",
     "cost": 40,
-    "buildtime": 3,
+    "buildTime": 3,
     "desc": "Monolithic structures that can hold many more things than we can think of.",
     "effects": [{"type": "terror", "effect": -30}, {"type": "pr", "effect": 3}],
     "prereq": [{ "type": "Technology", "code": "Inf4G"}],
@@ -209,7 +209,7 @@
     "code": "KI-8",
     "unitType": "MegaProject",
     "cost": 70,
-    "buildtime": 3,
+    "buildTime": 3,
     "desc": "Monolithic structures which attempt to bridge Space to the very Earth.",
     "effects": [{"type": "terror", "effect": -30}, {"type": "pr", "effect": 3}],
     "prereq": [{ "type": "Technology", "code": "Inf4G"}],
@@ -221,7 +221,7 @@
     "code": "KI-9",
     "unitType": "Base",
     "cost": 30,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Greg, your friendly neighborhood scientist, has signed up to work in your lab",
     "effects": [{"sciRate": 60}],
     "prereq": [{ "type": "Technology", "code": "Inf5B"}],
@@ -233,7 +233,7 @@
     "code": "KI-8",
     "unitType": "MegaProject",
     "cost": 100,
-    "buildtime": 3,
+    "buildTime": 3,
     "desc": "As you watch them, the particles are constantly changing where they are.",
     "effects": [{"type": "terror", "effect": -30}, {"type": "pr", "effect": 3}],
     "prereq": [{ "type": "Technology", "code": "Inf4G"}],
@@ -245,7 +245,7 @@
     "code": "KM-1",
     "unitType": "Civilian",
     "cost": 10,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "You are slowly helping make the world a safer place, a place free from communter nightmares and far fewer crashes.",
     "effects": [{"type": "terror", "effect": -10}],
     "prereq": [{ "type": "Technology", "code": "Mil2E"}],
@@ -257,7 +257,7 @@
     "code": "KM-2",
     "unitType": "Civilian",
     "cost": 15,
-    "buildtime": 2,
+    "buildTime": 2,
     "desc": "Robots programmed reminiscent of the 3 Laws are now integrated and optimized into your workforce.",
     "effects": [{"type": "terror", "effect": -10}, {"type": "$M", "effect": 1}],
     "prereq": [{ "type": "Technology", "code": "Mil3E"}],
